Add item explorer facade and component tests
