+++ frontend/src/test-setup.ts
@@ -1,4 +1,8 @@
 import '@angular/compiler';
+import { getTestBed } from '@angular/core/testing';
+import { BrowserTestingModule, platformBrowserTesting } from '@angular/platform-browser/testing';
+
+getTestBed().initTestEnvironment(BrowserTestingModule, platformBrowserTesting());
 
 type StorageMap = Map<string, string>;
